Update store service
- Complete store service.
- Add docker container : you can run the application in a container using the docker run command

+++ common/common-domain/src/main/java/com/inditex/domain/valueobject/BrandId.java
@@ -3,9 +3,9 @@ package com.inditex.domain.valueobject;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-public class BrandId extends BaseId<Double>{
+public class BrandId extends BaseId<Integer>{
 
-    public BrandId(Double value) {
+    public BrandId(Integer value) {
         super(value);
     }
 }
