+++ common/common-domain/src/main/java/com/inditex/domain/valueobject/ProductId.java
@@ -1,7 +1,7 @@
 package com.inditex.domain.valueobject;
 
-public class ProductId extends BaseId<Double> {
-    public ProductId(Double value) {
+public class ProductId extends BaseId<Long> {
+    public ProductId(Long value) {
         super(value);
     }
 }
